guide: Add entities for stable and unstable series
Sometimes referring to the exact version in the documentation is a bit
superfluous, and instead we just want to refer to general GnuCash release
series (2.2, 2.3, 2.4, etc.) in which something happened.

These two entity definitions refer to the current stable and unstable
releases of GnuCash. They must be updated in trunk and the current release
branch whenever a new series becomes active--for example, for the upcoming
2.4 (stable) and 2.5 (unstable) series.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash-docs/trunk@19700 57a11ea4-9604-0410-9ed3-97b8803252fd